fixed maven plugin stupid bug, new testdata for generateTest, renamed package from transform to provide
