+++ allure-report/allure-report-data/src/main/java/ru/yandex/qatools/allure/data/providers/XUnitDataProvider.java
@@ -1,4 +1,4 @@
-package ru.yandex.qatools.allure.data.transform;
+package ru.yandex.qatools.allure.data.providers;
 
 import ru.yandex.qatools.allure.data.AllureXUnit;
