+++ allure-report/allure-report-data/src/test/java/ru/yandex/qatools/allure/data/ReportGeneratorTest.java
@@ -12,16 +12,23 @@ import static org.junit.Assert.assertEquals;
  * @author Dmitry Baev charlie@yandex-team.ru
  *         Date: 21.10.13
  */
-public class AttachmentsTest {
+public class ReportGeneratorTest {
 
     @Rule
     public TemporaryFolder folder = new TemporaryFolder();
 
     @Test
+    public void generateWithEmptyInputTest() throws Exception {
+        File output = folder.newFolder();
+        AllureReportGenerator reportGenerator = new AllureReportGenerator();
+        reportGenerator.generate(output);
+    }
+
+    @Test
     public void attachmentsCountTest() throws Exception {
         assertEquals(
                 "Not all attachment files are found.",
-                3,
+                6,
                 AllureReportGenerator.getAttachmentsFiles(getTestDataDirs()).length
         );
     }
@@ -32,7 +39,7 @@ public class AttachmentsTest {
         AllureReportGenerator.copyAttachments(getTestDataDirs(), outputDirectory);
         assertEquals(
                 "Not all attachment files are copied.",
-                3,
+                6,
                 AllureReportGenerator.getAttachmentsFiles(outputDirectory).length
         );
     }
