+++ allure-report/allure-report-data/src/test/resources/origin/12354f9ce9437fd52af582f7e99d1892-testcase.json
@@ -1,0 +1,107 @@
+{
+  "uid" : "12354f9ce9437fd52af582f7e99d1892",
+  "title" : "Lot of steps test",
+  "time" : {
+    "start" : 1386599321626,
+    "stop" : 1386599321630,
+    "duration" : 4
+  },
+  "summary" : {
+    "attachments" : 0,
+    "steps" : 6
+  },
+  "failure" : {
+    "message" : "No file found.",
+    "stackTrace" : "java.io.FileNotFoundException: No file found.\n                    at my.company.steps.CommonSteps.makeThisTestBroken(CommonSteps.java:29)\n                    at my.company.TestWithSteps.lotOfStepsTest(TestWithSteps.java:34)\n                    at sun.reflect.NativeMethodAccessorImpl.invoke0(Native Method)\n                    at sun.reflect.NativeMethodAccessorImpl.invoke(NativeMethodAccessorImpl.java:57)\n                    at sun.reflect.DelegatingMethodAccessorImpl.invoke(DelegatingMethodAccessorImpl.java:43)\n                    at java.lang.reflect.Method.invoke(Method.java:601)\n                    at org.junit.runners.model.FrameworkMethod$1.runReflectiveCall(FrameworkMethod.java:47)\n                    at org.junit.internal.runners.model.ReflectiveCallable.run(ReflectiveCallable.java:12)\n                    at org.junit.runners.model.FrameworkMethod.invokeExplosively(FrameworkMethod.java:44)\n                    at org.junit.internal.runners.statements.InvokeMethod.evaluate(InvokeMethod.java:17)\n                    at org.junit.rules.TestWatcher$1.evaluate(TestWatcher.java:55)\n                    at org.junit.rules.RunRules.evaluate(RunRules.java:20)\n                    at org.junit.runners.ParentRunner.runLeaf(ParentRunner.java:271)\n                    at org.junit.runners.BlockJUnit4ClassRunner.runChild(BlockJUnit4ClassRunner.java:70)\n                    at org.junit.runners.BlockJUnit4ClassRunner.runChild(BlockJUnit4ClassRunner.java:50)\n                    at org.junit.runners.ParentRunner$3.run(ParentRunner.java:238)\n                    at org.junit.runners.ParentRunner$1.schedule(ParentRunner.java:63)\n                    at org.junit.runners.ParentRunner.runChildren(ParentRunner.java:236)\n                    at org.junit.runners.ParentRunner.access$000(ParentRunner.java:53)\n                    at org.junit.runners.ParentRunner$2.evaluate(ParentRunner.java:229)\n                    at org.junit.rules.TestWatcher$1.evaluate(TestWatcher.java:55)\n                    at org.junit.rules.RunRules.evaluate(RunRules.java:20)\n                    at org.junit.runners.ParentRunner.run(ParentRunner.java:309)\n                    at org.apache.maven.surefire.junit4.JUnit4TestSet.execute(JUnit4TestSet.java:53)\n                    at org.apache.maven.surefire.junit4.JUnit4Provider.executeTestSet(JUnit4Provider.java:123)\n                    at org.apache.maven.surefire.junit4.JUnit4Provider.invoke(JUnit4Provider.java:104)\n                    at sun.reflect.NativeMethodAccessorImpl.invoke0(Native Method)\n                    at sun.reflect.NativeMethodAccessorImpl.invoke(NativeMethodAccessorImpl.java:57)\n                    at sun.reflect.DelegatingMethodAccessorImpl.invoke(DelegatingMethodAccessorImpl.java:43)\n                    at java.lang.reflect.Method.invoke(Method.java:601)\n                    at org.apache.maven.surefire.util.ReflectionUtils.invokeMethodWithArray(ReflectionUtils.java:164)\n                    at org.apache.maven.surefire.booter.ProviderFactory$ProviderProxy.invoke(ProviderFactory.java:110)\n                    at org.apache.maven.surefire.booter.SurefireStarter.invokeProvider(SurefireStarter.java:175)\n                    at\n                    org.apache.maven.surefire.booter.SurefireStarter.runSuitesInProcessWhenForked(SurefireStarter.java:107)\n                    at org.apache.maven.surefire.booter.ForkedBooter.main(ForkedBooter.java:68)\n                "
+  },
+  "description" : "I have many steps\n\n                Parent Step\n                Last\n                MakeThisTestBroken\n            ",
+  "severity" : "BLOCKER",
+  "status" : "BROKEN",
+  "steps" : [ {
+    "title" : "Parent step",
+    "time" : {
+      "start" : 1386599321627,
+      "stop" : 1386599321628,
+      "duration" : 1
+    },
+    "summary" : {
+      "attachments" : 0,
+      "steps" : 3
+    },
+    "status" : "PASSED",
+    "attachments" : [ ],
+    "steps" : [ {
+      "title" : "\"First nested\" step",
+      "time" : {
+        "start" : 1386599321628,
+        "stop" : 1386599321628,
+        "duration" : 0
+      },
+      "summary" : {
+        "attachments" : 0,
+        "steps" : 0
+      },
+      "status" : "PASSED",
+      "attachments" : [ ],
+      "steps" : [ ]
+    }, {
+      "title" : "\"Second nested\" step",
+      "time" : {
+        "start" : 1386599321628,
+        "stop" : 1386599321628,
+        "duration" : 0
+      },
+      "summary" : {
+        "attachments" : 0,
+        "steps" : 0
+      },
+      "status" : "PASSED",
+      "attachments" : [ ],
+      "steps" : [ ]
+    }, {
+      "title" : "\"Third nested\" step",
+      "time" : {
+        "start" : 1386599321628,
+        "stop" : 1386599321628,
+        "duration" : 0
+      },
+      "summary" : {
+        "attachments" : 0,
+        "steps" : 0
+      },
+      "status" : "PASSED",
+      "attachments" : [ ],
+      "steps" : [ ]
+    } ]
+  }, {
+    "title" : "\"Last\" step",
+    "time" : {
+      "start" : 1386599321628,
+      "stop" : 1386599321628,
+      "duration" : 0
+    },
+    "summary" : {
+      "attachments" : 0,
+      "steps" : 0
+    },
+    "status" : "PASSED",
+    "attachments" : [ ],
+    "steps" : [ ]
+  }, {
+    "title" : "Make this test broken",
+    "time" : {
+      "start" : 1386599321629,
+      "stop" : 1386599321629,
+      "duration" : 0
+    },
+    "summary" : {
+      "attachments" : 0,
+      "steps" : 0
+    },
+    "status" : "BROKEN",
+    "attachments" : [ ],
+    "steps" : [ ]
+  } ],
+  "attachments" : [ ],
+  "labels" : [ ]
+}
