+++ allure-report/allure-report-data/src/test/resources/origin/53064e37-3632-495b-87fd-4a69219b2e02-attachment.xml
@@ -1,0 +1,6 @@
+<root>
+    <a href="some"/>
+    <b>
+        <c/>
+    </b>
+</root>
