+++ allure-report/allure-report-data/src/test/resources/origin/5f73d429-dadb-4fba-9a24-a3ed88563ebf-attachment.txt
@@ -1,0 +1,17 @@
+Сегодня к Яндекс.Маркету подключено больше 13 тысяч магазинов, а посещают его больше 17 миллионов человек в месяц (по данным TNS). В скором времени на сервисе планируются серьёзные изменения: пользователи получат возможность не только выбирать товар на Яндекс.Маркете, но и оформлять заказ прямо на сервисе — без перехода на сайт магазина. А магазины получат новую модель взаимодействия с Маркетом.
+
+
+
+Новая модель — с оплатой за действие (Cost Per Action, CPA). До сих пор партнёры оплачивали переходы на свои сайты (Cost Per Click, CPC). Мы запускаем размещение по CPA в дополнение к ней. 
+
+Не всеми магазинами, подключенными к нашему сервису, люди пользуются так же часто, как Маркетом. Его интерфейс им привычен и понятен. Мы предусмотрели для него разные представления в зависимости от того, что человек собирается искать. На Маркете неодинаковый интерфейс для выбора сумок и холодильников, но зато одна корзина, в которую можно сложить разные типы товаров и разом их заказать. 
+
+Модель, в которой оформить всё можно будет прямо на Маркете, упростит людям процесс покупки и поможет магазинам увеличить число заказов. Последние должны будут оплачивать только 1% от суммы, причем независимо от категории товара. Это самая низкая на рынке комиссия с заказа для интернет-магазинов. Подключаться к Маркету по модели CPA можно уже сейчас — через специальный API. 
+
+Яндекс.Маркет не станет конкурентом магазинам, а продолжит быть площадкой для их честной конкуренции между собой. Благодаря тому, что мы агрегируем предложения сторонних магазинов, разрабатываем технологии для удобного выбора и сравнения товаров, реализовали систему рейтингов и отзывов, сервис поспособствовал развитию электронной коммерции в России, Украине и других странах.
+
+
+
+Переход на новую модель даст владельцам магазинов новое преимущество: они смогут платить за реальный результат и проще считать свои расходы на привлечение клиентов. Подключиться к Маркету по модели CPA могут российские и украинские магазины, которые не менее трех месяцев работают с сервисом и имеют рейтинг не ниже трех звезд. 
+
+Возможность совершать покупки без перехода на сайты магазинов появится ближе к концу этого года. Но Маркет — это сервис, который объединяет в себе пользовательскую и партнерскую части. Поэтому сейчас мы анонсируем новые возможности, чтобы магазины начали интеграцию с «Покупкой на Маркете» и мы вместе смогли предоставить пользователям качественный сервис.
