--- allure-report/allure-report-data/src/test/resources/origin/90151ae515bb91b1efab7d57a08d15a9-testcase.json
@@ -1,20 +1,0 @@
-{
-  "uid" : "90151ae515bb91b1efab7d57a08d15a9",
-  "title" : "Parameterized test[1]",
-  "time" : {
-    "start" : 1386463101600,
-    "stop" : 1386463101600,
-    "duration" : 0
-  },
-  "summary" : {
-    "attachments" : 0,
-    "steps" : 0
-  },
-  "failure" : null,
-  "description" : null,
-  "severity" : "NORMAL",
-  "status" : "PASSED",
-  "steps" : [ ],
-  "attachments" : [ ],
-  "labels" : [ ]
-}
