--- allure-report/allure-report-data/src/test/resources/origin/9051a2dac33abb1b4c43a0e72c847254-testcase.json
@@ -1,36 +1,0 @@
-{
-  "uid" : "9051a2dac33abb1b4c43a0e72c847254",
-  "title" : "Simple test with attachments",
-  "time" : {
-    "start" : 1386463101857,
-    "stop" : 1386463101875,
-    "duration" : 18
-  },
-  "summary" : {
-    "attachments" : 1,
-    "steps" : 1
-  },
-  "failure" : null,
-  "description" : null,
-  "severity" : "NORMAL",
-  "status" : "PASSED",
-  "steps" : [ ],
-  "attachments" : [ {
-    "title" : "makeAttach",
-    "source" : "a9852d34-96aa-43f1-b0e1-ee69d104900e-attachment.txt",
-    "type" : "TXT"
-  } ],
-  "labels" : [ {
-    "name" : "Story",
-    "value" : "MyAmazingStory"
-  }, {
-    "name" : "Feature",
-    "value" : "MySecondFeature"
-  }, {
-    "name" : "Story",
-    "value" : "MyHorrorStory"
-  }, {
-    "name" : "Feature",
-    "value" : "MySecondFeature"
-  } ]
-}
