--- allure-report/allure-report-data/src/test/resources/origin/a9852d34-96aa-43f1-b0e1-ee69d104900e-attachment.txt
@@ -1,1 +1,0 @@
-yeah, 2 is 2
