+++ allure-report/allure-report-data/src/test/resources/origin/behavior.json
@@ -1,96 +1,3 @@
 {
-  "features" : [ {
-    "title" : "MyFirstFeature",
-    "statistic" : {
-      "total" : 1,
-      "passed" : 1,
-      "skipped" : 0,
-      "failed" : 0,
-      "broken" : 0
-    },
-    "stories" : [ {
-      "uid" : "6ec8f5740e69d8fb780448440e6bf6d9",
-      "title" : "MyFirstStory",
-      "statistic" : {
-        "total" : 1,
-        "passed" : 1,
-        "skipped" : 0,
-        "failed" : 0,
-        "broken" : 0
-      },
-      "testCases" : [ {
-        "uid" : "d6fdfd42ac420296da726828a34ab5fc",
-        "title" : "Simple test with steps",
-        "time" : {
-          "start" : 1386463101842,
-          "stop" : 1386463101855,
-          "duration" : 13
-        },
-        "severity" : "NORMAL",
-        "status" : "PASSED"
-      } ]
-    } ]
-  }, {
-    "title" : "MySecondFeature",
-    "statistic" : {
-      "total" : 3,
-      "passed" : 3,
-      "skipped" : 0,
-      "failed" : 0,
-      "broken" : 0
-    },
-    "stories" : [ {
-      "uid" : "df2a5a56c512c5fa23f68723fc50a09f",
-      "title" : "MyAmazingStory",
-      "statistic" : {
-        "total" : 2,
-        "passed" : 2,
-        "skipped" : 0,
-        "failed" : 0,
-        "broken" : 0
-      },
-      "testCases" : [ {
-        "uid" : "eb8e4ca2761d8f23962df89f0c98e722",
-        "title" : "Simple test",
-        "time" : {
-          "start" : 1386463101855,
-          "stop" : 1386463101857,
-          "duration" : 2
-        },
-        "severity" : "NORMAL",
-        "status" : "PASSED"
-      }, {
-        "uid" : "9051a2dac33abb1b4c43a0e72c847254",
-        "title" : "Simple test with attachments",
-        "time" : {
-          "start" : 1386463101857,
-          "stop" : 1386463101875,
-          "duration" : 18
-        },
-        "severity" : "NORMAL",
-        "status" : "PASSED"
-      } ]
-    }, {
-      "uid" : "a69669bb40a70dbfe9c3c21bd7590326",
-      "title" : "MyHorrorStory",
-      "statistic" : {
-        "total" : 1,
-        "passed" : 1,
-        "skipped" : 0,
-        "failed" : 0,
-        "broken" : 0
-      },
-      "testCases" : [ {
-        "uid" : "9051a2dac33abb1b4c43a0e72c847254",
-        "title" : "Simple test with attachments",
-        "time" : {
-          "start" : 1386463101857,
-          "stop" : 1386463101875,
-          "duration" : 18
-        },
-        "severity" : "NORMAL",
-        "status" : "PASSED"
-      } ]
-    } ]
-  } ]
+  "features" : [ ]
 }
