--- allure-report/allure-report-data/src/test/resources/origin/eb8e4ca2761d8f23962df89f0c98e722-testcase.json
@@ -1,26 +1,0 @@
-{
-  "uid" : "eb8e4ca2761d8f23962df89f0c98e722",
-  "title" : "Simple test",
-  "time" : {
-    "start" : 1386463101855,
-    "stop" : 1386463101857,
-    "duration" : 2
-  },
-  "summary" : {
-    "attachments" : 1,
-    "steps" : 1
-  },
-  "failure" : null,
-  "description" : null,
-  "severity" : "NORMAL",
-  "status" : "PASSED",
-  "steps" : [ ],
-  "attachments" : [ ],
-  "labels" : [ {
-    "name" : "Story",
-    "value" : "MyAmazingStory"
-  }, {
-    "name" : "Feature",
-    "value" : "MySecondFeature"
-  } ]
-}
