+++ allure-report/allure-report-data/src/test/resources/origin/f2afdc4d-8b7a-48b9-9e7d-bc8f8d3e6445-attachment.json
@@ -1,0 +1,24 @@
+{
+    "uid": "9a86126cdcb922afbf4f950b533f49045b595c826c0b44938a3421ae58fce017",
+    "parentUid": "c51469d88be9f49cf313cabc267314e84dd67d1824608e01a03a8cc0bfe25a88",
+    "name": "failedTestWithoutSteps",
+    "title": "Failed test without steps",
+    "description": null,
+    "classname": "my.company.TestWithoutSteps",
+    "failure": {
+        "message": "5 is not 2 + 2\nExpected: is <4>\n     but: was <5>",
+        "stackTrace": "java.lang.AssertionError: 5 is not 2 + 2\nExpected: is <4>\n     but: was <5>\n\tat org.hamcrest.MatcherAssert.assertThat(MatcherAssert.java:20)\n\tat org.junit.Assert.assertThat(Assert.java:865)\n\tat my.company.TestWithoutSteps.failedTestWithoutSteps(TestWithoutSteps.java:25)\n\tat sun.reflect.NativeMethodAccessorImpl.invoke0(Native Method)\n\tat sun.reflect.NativeMethodAccessorImpl.invoke(NativeMethodAccessorImpl.java:57)\n\tat sun.reflect.DelegatingMethodAccessorImpl.invoke(DelegatingMethodAccessorImpl.java:43)\n\tat java.lang.reflect.Method.invoke(Method.java:601)\n\tat org.junit.runners.model.FrameworkMethod$1.runReflectiveCall(FrameworkMethod.java:47)\n\tat org.junit.internal.runners.model.ReflectiveCallable.run(ReflectiveCallable.java:12)\n\tat org.junit.runners.model.FrameworkMethod.invokeExplosively(FrameworkMethod.java:44)\n\tat org.junit.internal.runners.statements.InvokeMethod.evaluate(InvokeMethod.java:17)\n\tat org.junit.rules.TestWatcher$1.evaluate(TestWatcher.java:55)\n\tat org.junit.rules.RunRules.evaluate(RunRules.java:20)\n\tat org.junit.runners.ParentRunner.runLeaf(ParentRunner.java:271)\n\tat org.junit.runners.BlockJUnit4ClassRunner.runChild(BlockJUnit4ClassRunner.java:70)\n\tat org.junit.runners.BlockJUnit4ClassRunner.runChild(BlockJUnit4ClassRunner.java:50)\n\tat org.junit.runners.ParentRunner$3.run(ParentRunner.java:238)\n\tat org.junit.runners.ParentRunner$1.schedule(ParentRunner.java:63)\n\tat org.junit.runners.ParentRunner.runChildren(ParentRunner.java:236)\n\tat org.junit.runners.ParentRunner.access$000(ParentRunner.java:53)\n\tat org.junit.runners.ParentRunner$2.evaluate(ParentRunner.java:229)\n\tat org.junit.rules.TestWatcher$1.evaluate(TestWatcher.java:55)\n\tat org.junit.rules.RunRules.evaluate(RunRules.java:20)\n\tat org.junit.runners.ParentRunner.run(ParentRunner.java:309)\n\tat org.apache.maven.surefire.junit4.JUnit4Provider.execute(JUnit4Provider.java:264)\n\tat org.apache.maven.surefire.junit4.JUnit4Provider.executeTestSet(JUnit4Provider.java:153)\n\tat org.apache.maven.surefire.junit4.JUnit4Provider.invoke(JUnit4Provider.java:124)\n\tat org.apache.maven.surefire.booter.ForkedBooter.invokeProviderInSameClassLoader(ForkedBooter.java:200)\n\tat org.apache.maven.surefire.booter.ForkedBooter.runSuitesInProcess(ForkedBooter.java:153)\n\tat org.apache.maven.surefire.booter.ForkedBooter.main(ForkedBooter.java:103)\n"
+    },
+    "tags": [],
+    "steps": [],
+    "attachments": [],
+    "storyInfo": null,
+    "featureInfo": null,
+    "start": 1381235848964,
+    "stop": 1381235848972,
+    "duration": 8,
+    "stepsCount": 0,
+    "attachmentsCount": 0,
+    "status": "FAILED",
+    "severity": "NORMAL"
+}
