+++ allure-report/allure-report-data/src/test/resources/origin/xunit.json
@@ -1,94 +1,46 @@
 {
   "time" : {
-    "start" : 1386463101570,
-    "stop" : 1386463101875,
-    "duration" : 305
+    "start" : 1386599321528,
+    "stop" : 1386599321631,
+    "duration" : 103
   },
   "testSuites" : [ {
-    "uid" : "f8000b87eee30d3906f763adfa621f71",
-    "title" : "Simple test",
+    "uid" : "93411df1130a181741c15487a635862b",
+    "title" : "Test with steps",
     "time" : {
-      "start" : 1386463101840,
-      "stop" : 1386463101875,
-      "duration" : 35
-    },
-    "statistic" : {
-      "total" : 3,
-      "passed" : 3,
-      "skipped" : 0,
-      "failed" : 0,
-      "broken" : 0
-    },
-    "description" : null,
-    "classname" : "my.company.tests.SimpleTest",
-    "testCases" : [ {
-      "uid" : "d6fdfd42ac420296da726828a34ab5fc",
-      "title" : "Simple test with steps",
-      "time" : {
-        "start" : 1386463101842,
-        "stop" : 1386463101855,
-        "duration" : 13
-      },
-      "severity" : "NORMAL",
-      "status" : "PASSED"
-    }, {
-      "uid" : "eb8e4ca2761d8f23962df89f0c98e722",
-      "title" : "Simple test",
-      "time" : {
-        "start" : 1386463101855,
-        "stop" : 1386463101857,
-        "duration" : 2
-      },
-      "severity" : "NORMAL",
-      "status" : "PASSED"
-    }, {
-      "uid" : "9051a2dac33abb1b4c43a0e72c847254",
-      "title" : "Simple test with attachments",
-      "time" : {
-        "start" : 1386463101857,
-        "stop" : 1386463101875,
-        "duration" : 18
-      },
-      "severity" : "NORMAL",
-      "status" : "PASSED"
-    } ]
-  }, {
-    "uid" : "ecb454d16dc38def6ec1ff6cc6a44159",
-    "title" : "Parameterized test",
-    "time" : {
-      "start" : 1386463101570,
-      "stop" : 1386463101600,
-      "duration" : 30
+      "start" : 1386599321528,
+      "stop" : 1386599321631,
+      "duration" : 103
     },
     "statistic" : {
       "total" : 2,
-      "passed" : 2,
+      "passed" : 0,
       "skipped" : 0,
-      "failed" : 0,
-      "broken" : 0
+      "failed" : 1,
+      "broken" : 1
     },
-    "description" : null,
-    "classname" : "my.company.tests.ParameterizedTest",
+    "description" : "Testsuite has testcases with many steps\n        and\n        many lines in description\n    ",
+    "classname" : "my.company.TestWithSteps",
     "testCases" : [ {
-      "uid" : "1155fcc5e23d3e4358cc9a811acd943c",
-      "title" : "Parameterized test[0]",
+      "uid" : "9076213cb0b87f92bb7bb85fc5399da3",
+      "title" : "Attachments test",
       "time" : {
-        "start" : 1386463101587,
-        "stop" : 1386463101595,
-        "duration" : 8
+        "start" : 1386599321574,
+        "stop" : 1386599321620,
+        "duration" : 46
       },
-      "severity" : "NORMAL",
-      "status" : "PASSED"
+      "severity" : "CRITICAL",
+      "status" : "FAILED"
     }, {
-      "uid" : "90151ae515bb91b1efab7d57a08d15a9",
-      "title" : "Parameterized test[1]",
+      "uid" : "12354f9ce9437fd52af582f7e99d1892",
+      "title" : "Lot of steps test",
       "time" : {
-        "start" : 1386463101600,
-        "stop" : 1386463101600,
-        "duration" : 0
+        "start" : 1386599321626,
+        "stop" : 1386599321630,
+        "duration" : 4
       },
-      "severity" : "NORMAL",
-      "status" : "PASSED"
+      "severity" : "BLOCKER",
+      "status" : "BROKEN"
     } ]
   } ]
 }
