--- allure-report/allure-report-data/src/test/resources/testdata3/05c98d9b-0b25-4a23-bcfc-a8d276535473-testsuite.xml
@@ -1,38 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" standalone="yes"?>
-<ns2:test-suite xmlns:ns2="urn:model.allure.qatools.yandex.ru" start="1386463101840" stop="1386463101875">
-    <title>Simple test</title>
-    <classname>my.company.tests.SimpleTest</classname>
-    <test-cases>
-        <test-case start="1386463101842" stop="1386463101855" status="passed" severity="normal">
-            <title>Simple test with steps</title>
-            <steps>
-                <step start="1386463101854" stop="1386463101855" status="passed">
-                    <title>checkThat2is2</title>
-                </step>
-            </steps>
-            <labels>
-                <label name="Story" value="MyFirstStory"/>
-                <label name="Feature" value="MyFirstFeature"/>
-            </labels>
-        </test-case>
-        <test-case start="1386463101855" stop="1386463101857" status="passed" severity="normal">
-            <title>Simple test</title>
-            <labels>
-                <label name="Story" value="MyAmazingStory"/>
-                <label name="Feature" value="MySecondFeature"/>
-            </labels>
-        </test-case>
-        <test-case start="1386463101857" stop="1386463101875" status="passed" severity="normal">
-            <title>Simple test with attachments</title>
-            <attachments>
-                <attachment title="makeAttach" source="a9852d34-96aa-43f1-b0e1-ee69d104900e-attachment.txt" type="txt"/>
-            </attachments>
-            <labels>
-                <label name="Story" value="MyAmazingStory"/>
-                <label name="Feature" value="MySecondFeature"/>
-                <label name="Story" value="MyHorrorStory"/>
-                <label name="Feature" value="MySecondFeature"/>
-            </labels>
-        </test-case>
-    </test-cases>
-</ns2:test-suite>
