--- allure-report/allure-report-data/src/test/resources/testdata3/0b6eebcb-bc3e-464e-ac5f-79fd92df8d0a-testsuite.xml
@@ -1,13 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" standalone="yes"?>
-<ns2:test-suite xmlns:ns2="urn:model.allure.qatools.yandex.ru" start="1386463101570" stop="1386463101600">
-    <title>Parameterized test</title>
-    <classname>my.company.tests.ParameterizedTest</classname>
-    <test-cases>
-        <test-case start="1386463101587" stop="1386463101595" status="passed" severity="normal">
-            <title>Parameterized test[0]</title>
-        </test-case>
-        <test-case start="1386463101600" stop="1386463101600" status="passed" severity="normal">
-            <title>Parameterized test[1]</title>
-        </test-case>
-    </test-cases>
-</ns2:test-suite>
